Make radix_key_codec part of the public API
