+++ test/rocprim/test_block_sort.hpp
@@ -1,6 +1,6 @@
 // MIT License
 //
-// Copyright (c) 2017-2023 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2024 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -22,9 +22,9 @@
 
 #include "test_utils_sort_comparator.hpp"
 
-block_sort_test_suite_type_def(suite_name, name_suffix)
+block_sort_test_suite_type_def(suite_name, name_suffix);
 
-    typed_test_suite_def(suite_name, name_suffix, block_params);
+typed_test_suite_def(suite_name, name_suffix, block_params);
 
 // using header guards for these test functions because this file is included multiple times:
 // once for the integrals test suite and once for the floating point test suite.
